fix(ingest): repair ingest-v4l2, lint rtsp-gstreamer; gate both in CI
Two more off-by-default features that no CI job compiled (surfaced while
hardening the optional-feature gates):

ingest-v4l2 was structurally broken (32 errors) by un-migrated v4l 0.14 /
ouroboros 0.18 bumps:
  - v4l 0.14 MmapStream dropped its device type param: MmapStream<'this>.
  - with_buffers takes &Device, so the self_referencing field borrows the
    device immutably (#[borrows(device)]).
  - ouroboros 0.18 fallible builds use the generated *TryBuilder* struct
    (DeviceV4l2StateTryBuilder), not Builder::try_build.
  - plus 1.95 clippy: checked_div, is_multiple_of, is_some, inspect_err,
    and removal of an orphaned dead helper. Synthetic (stub://) unit tests
    pass; the device path can't be exercised without hardware.

rtsp-gstreamer compiled fine but had two never-linted unnecessary u32->u32
casts in sample_to_pixels; removed.

CI: add ingest-v4l2 to the optional-features job (the v4l crate builds and
links without a system library), and add an ingest-gstreamer job that
installs the GStreamer dev libraries and clippy-gates rtsp-gstreamer.
Verified all green with cargo +1.95.0 clippy --all-targets -D warnings.

+++ src/ingest/rtsp.rs
@@ -435,8 +435,8 @@ fn sample_to_pixels(sample: &gstreamer::Sample) -> Result<(Vec<u8>, u32, u32)> {
     let info =
         gstreamer_video::VideoInfo::from_caps(caps).context("parse RTSP caps as video info")?;
 
-    let width = info.width() as u32;
-    let height = info.height() as u32;
+    let width = info.width();
+    let height = info.height();
     let row_bytes = (width as usize) * 3;
     let stride = info.stride()[0] as usize;
